Created a class deffinition and partially implemented a definition
Not completely set on how things are totally fine if things get changed.
Havent tested the member functions yet but the code was tested when in
the stegoTest.cpp before turning into memmber functions. The contents of
int main have what could be used for the begining of a decode function.